+++ bolshakovPetr/app/(tabs)/explore.tsx
